// packages/docusaurus-sylk-plugin/src/sylk/protos/sylk/SylkEnum/v2/SylkEnum.d.ts
export interface SylkEnum {
    type: string;
    kind: string;
    description: string;
    values: SylkEnumValue[];
    name: string;
    uri: string;
    fullName: string;
    tag: string;
}