// packages/docusaurus-sylk-plugin/src/sylk/protos/sylk/SylkProject/v1/SylkProject.d.ts
export interface SylkProject {
    description: string;
    javaPackage: string;
    goPackage: string;
    name: string;
    uri: string;
    clients: SylkClient[];
    server?: SylkServer;
    packageName: string;
    extensions: Any[];
}