// packages/docusaurus-sylk-plugin/src/sylk/protos/sylk/SylkService/v2/SylkService.d.ts
export interface SylkService {
    dependencies: string[];
    description: string;
    uri: string;
    name: string;
    fullName: string;
    type: string;
    methods: SylkMethod[];
    extensions: {
        [key: string]: {
            [key: string]: any;
        };
    };
    tag: string;
}